+++ packages/dash-core-components/src/components/Slider.react.js
@@ -18,7 +18,7 @@ export default class Slider extends Component {
     }
 
     render() {
-        const {id, setProps, updatemode, loading_state} = this.props;
+        const {id, loading_state, setProps, updatemode, vertical} = this.props;
         const {value} = this.state;
         return (
             <div
@@ -26,6 +26,7 @@ export default class Slider extends Component {
                 data-dash-is-loading={
                     (loading_state && loading_state.is_loading) || undefined
                 }
+                style={vertical ? {height: '100%'} : {}}
             >
                 <ReactSlider
                     onChange={value => {